FE: Applied new startup on app
